mudei o delete user para pegar o id a partir do token

+++ src/controllers/userController.js
@@ -50,14 +50,12 @@ module.exports = {
   },
 
   async delete(request, response) {
-    const { id } = request.params;
+    try {
+      await User.destroy({ where: { id: request.userId } });
 
-    if (request.userId !== Number(id)) {
-      return sendError(response, status.FORBIDDEN);
+      return response.sendStatus(status.OK);
+    } catch (e) {
+      return sendError(response, status.SERVER_ERROR);
     }
-
-    await User.destroy({ where: { id } });
-
-    return response.sendStatus(200);
   },
 };
